requirements.txt
streamlit>=1.34.0
google-generativeai>=0.7.0
python-dotenv>=1.0.1
manim>=0.18.0
moviepy>=1.0.3
gTTS>=2.5.0
# (Manim commonly uses these; usually auto-installed)
numpy
pillow
firebase-admin>=6.4.0
google-cloud-firestore>=2.14.0
streamlit-authenticator>=0.2.3


